app/auth_service/requirements.txt
alembic==1.14.0
fastapi==0.115.6
httpx==0.28.1
passlib==1.7.4
pydantic==2.10.4
pydantic_settings==2.7.0
PyJWT==2.10.1
pytest==8.3.4
python_json_logger==3.2.1
SQLAlchemy==2.0.36
pytest_httpx==0.35.0
python_json_logger==3.2.1
asyncpg==0.29.0
gunicorn==23.0.0
uvicorn==0.30.1
python-multipart==0.0.9
pydantic[email]
taskiq==0.11.7
fastapi-cache2==0.2.2
taskiq_redis==1.0.0
botocore==1.34.131
aiobotocore==2.13.1
aio_pika==9.4.3